requirements.txt
"""
TORIS AI - Requirements File
Implements pinned dependencies with version constraints
"""
# Core dependencies
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.23.0
httpx>=0.24.0,<0.25.0
pydantic>=1.10.7,<2.0.0
loguru>=0.7.0,<0.8.0
python-dotenv>=1.0.0,<1.1.0

# API security and performance
slowapi>=0.1.7,<0.2.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0

# Memory and vector storage
chromadb>=0.4.6,<0.5.0
sentence-transformers>=2.2.2,<2.3.0

# LLM and tools
ollama>=0.1.0,<0.2.0
langchain>=0.0.267,<0.1.0
beautifulsoup4>=4.12.2,<4.13.0
requests>=2.31.0,<2.32.0
pillow>=10.0.0,<10.1.0

# UI
gradio>=3.36.1,<3.37.0

# Code execution and sandboxing
docker>=6.1.2,<6.2.0

# Utilities
numpy>=1.24.3,<1.25.0
pandas>=2.0.2,<2.1.0
tqdm>=4.65.0,<4.66.0
